requirements.txt
anthropic>=0.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
streamlit>=1.40.0
pymupdf>=1.24.0
numpy>=1.26.0
